test: isolate recovered message mapper
